Adds clocking time for writing in db_runner

+++ src/tmpdb/fluid_options.hpp
@@ -18,9 +18,9 @@ typedef enum {ENTRIES = 0, LEVELS = 1} bulk_load_type;
 class FluidOptions
 {
 public:
-    double size_ratio = 2;              //> (T)
-    double lower_level_run_max = 1;     //> (K)
-    double largest_level_run_max = 1;   //> (Z)
+    int size_ratio = 2;              //> (T)
+    int lower_level_run_max = 1;     //> (K)
+    int largest_level_run_max = 1;   //> (Z)
     size_t buffer_size = 1048576;       //> bytes (B)
     size_t entry_size = 8192;           //> bytes (E)
     double bits_per_element = 5.0;      //> bits per element per bloom filter at all levels (h)
